requirements.txt
langchain
langchain_experimental
langchain_openai
pandas
python-dotenv
tabulate
gradio
